refactor: extract shared modelKey/parseModelKey to src/lib/model-keys
